add separate class for setup client
